+++ src/sass/_certificate_add.scss
@@ -4,16 +4,4 @@
     justify-content: space-between;
     overflow-x: auto;
   }
-
-  .identity-trust-token-command-wrapper code {
-    width: 100%;
-
-    .command-wrapper {
-      display: flex;
-
-      .command {
-        flex-grow: 1;
-      }
-    }
-  }
 }
