+++ src/sass/_permission_confirm_modal.scss
@@ -48,13 +48,3 @@
     margin-top: $spv--large;
   }
 }
-
-.token-code-block {
-  background-color: rgb(0 0 0 / 5%);
-  margin-bottom: 1.5rem;
-  padding: 0.5rem 1rem;
-
-  code {
-    background-color: transparent;
-  }
-}
